requirements.txt
absl-py==2.3.1
ai2-olmo==0.6.0
ai2-olmo-core==0.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
appnope==0.1.4
asttokens==3.0.0
attrs==25.4.0
babel==2.17.0
backcall==0.2.0
beautifulsoup4==4.14.2
bleach==6.3.0
boto3==1.40.71
botocore==1.40.71
bs4==0.0.2
cached_path==1.8.0
cachetools==6.2.1
certifi==2025.11.12
chardet==5.2.0
charset-normalizer==2.1.1
click==8.3.0
contourpy==1.3.3
cycler==0.12.1
datasets==4.4.1
decorator==5.2.1
defusedxml==0.7.1
dill==0.4.0
diskcache==5.6.3
docopt==0.6.2
docx2txt==0.9
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.20.0
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.10.0
gensim==4.4.0
google-api-core==2.28.1
google-auth==2.41.1
google-auth-oauthlib==1.2.3
google-cloud-core==2.5.0
google-cloud-storage==2.19.0
google-crc32c==1.8.0rc3
google-measurement-protocol==1.1.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
intel-cmplr-lib-ur==2025.3.1
intel-openmp==2025.3.1
ipython==8.12.3
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyterlab_pygments==0.3.0
kiwisolver==1.4.9
llama_cpp_python==0.3.16
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
mdurl==0.1.2
mistune==3.1.4
mkl==2025.3.0
mkl-service==2.6.0
mkl_fft @ file:///home/task_176159258761192/conda-bld/mkl_fft_1761592901799/work
mkl_random @ file:///home/task_176159293348456/conda-bld/mkl_random_1761592947335/work
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.18
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
networkx==3.5
nltk==3.9.2
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
oauthlib==3.3.1
omegaconf==2.3.0
onemkl-license==2025.3.0
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
pexpect==4.9.0
pickleshare==0.7.5
pillow==12.0.0
pipreqs==0.5.0
platformdirs==4.5.0
prices==1.1.1
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.0
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycocoevalcap==1.2
pycocotools==2.0.10
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.28.0
rsa==4.9.1
s3transfer==0.14.0
safetensors==0.6.2
scikit-learn==1.7.2
scikit-video==1.1.11
scipy==1.16.3
seaborn==0.13.2
shellingham==1.5.4
six==1.17.0
smart_open==7.5.0
sniffio==1.3.1
SoccerNet==0.1.62
soupsieve==2.8
stack-data==0.6.3
sympy==1.14.0
tbb==2022.3.0
tcmlib==1.4.1
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
thop==0.1.1.post2209072238
threadpoolctl==3.6.0
tinycss2==1.4.0
tokenizers==0.22.1
torch==2.9.0
torchvision==0.24.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.1
triton==3.5.0
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
umf==1.0.2
urllib3==2.5.0
wcwidth==0.2.14
webencodings==0.5.1
Werkzeug==3.1.3
wrapt==2.0.1
xxhash==3.6.0
yarg==0.1.10
yarl==1.22.0
zipp==3.23.0
